+++ src/main.jsx
@@ -1,10 +1,10 @@
-import React from "react"
-import ReactDOM from "react-dom/client"
-import App from "./App.jsx"
-import "./index.css"
-import("preline")
-import { BrowserRouter } from "react-router-dom"
-import ApiState from "./Context/api/apiState.jsx"
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App.jsx";
+import "./assets/styles/index.css";
+import("preline");
+import { BrowserRouter } from "react-router-dom";
+import ApiState from "./Context/api/apiState.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
@@ -14,4 +14,4 @@ ReactDOM.createRoot(document.getElementById("root")).render(
       </BrowserRouter>
     </ApiState>
   </React.StrictMode>
-)
+);
