+++ src/pages/Landing/LandingPage.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Hero from "../components/Hero";
-import Blogs from "../components/Blogs";
-import Features from "../components/Features";
-import FAQ from "../components/FAQ";
+import Hero from "./components/Hero";
+import Blogs from "./components/Blogs";
+import Features from "./components/Features";
+import FAQ from "./components/FAQ";
 // import Newsletter from "../components/Newsletter";
-import Footer from "../components/Footer";
+import Footer from "./components/Footer";
 
 const LandingPage = () => {
   return (
