+++ src/pages/Landing/components/FAQ.jsx
